add another conversion

+++ lib/image_processor.py
@@ -176,7 +176,7 @@ class ImageProcessor:
 
         # recalculate draw image with new predicted size
         if not line_break:
-            font_size *= original_img.size[0] / new_img.size[0]
+            font_size *= int(original_img.size[0] / new_img.size[0])
             font = ImageFont.truetype(font_type, size=font_size)
             new_img = Image.new('RGBA', (x + padding * 2, new_img_height + padding), color=(0,0,0,0))
             draw = ImageDraw.Draw(new_img)
